Move send_command to pywundasmart
And remove unused code
